Better error messages

+++ src/grib_dumper_class_wmo.c
@@ -367,7 +367,7 @@ static void dump_bytes(grib_dumper* d, grib_accessor* a, const char* comment)
         if (size == 0)
             fprintf(self->dumper.out, "}\n");
         else
-            fprintf(self->dumper.out, " *** ERR cannot malloc(%ld) }\n", (long)size);
+            fprintf(self->dumper.out, " *** ERR cannot malloc(%zu) }\n", size);
         return;
     }
 
@@ -465,7 +465,7 @@ static void dump_values(grib_dumper* d, grib_accessor* a)
         if (size == 0)
             fprintf(self->dumper.out, "}\n");
         else
-            fprintf(self->dumper.out, " *** ERR cannot malloc(%ld) }\n", (long)size);
+            fprintf(self->dumper.out, " *** ERR cannot malloc(%zu) }\n", size);
         return;
     }
